[FIX] Install Chrome fixed version to avoid sudden errors
Odoo's approach with Runbots is to install a fixed version of Chrome and
only when a new version is tested across all the supported code base
they bump it.

That way, they avoid sudden errors that would make the whole CIs
collapse on tour tests.

We can follow the guideline of Odoo's fixed versions for a minimum
guarantee.